--- blocks/alternative_mutators/plus_minus/package.json
@@ -1,26 +1,0 @@
-{
-    "name": "@blockly/plugin-plus-minus-mutators",
-    "version": "0.0.0",
-    "private": true,
-    "scripts": {
-        "build": "webpack --display-error-details --env.mode=development",
-        "test": "mocha --ui tdd"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.8.6",
-        "@babel/preset-env": "^7.8.6",
-        "babel-loader": "^8.0.6",
-        "blockly": "^3.20200123.1",
-        "chai": "^4.2.0",
-        "eslint": "^6.8.0",
-        "eslint-plugin-jsdoc": "^22.0.0",
-        "mocha": "^7.1.0",
-        "sinon": "7.5.0",
-        "webpack": "^4.41.6",
-        "webpack-cli": "^3.3.11",
-        "webpack-dev-server": "^3.10.3"
-    },
-    "dependencies": {
-        "blockly": "^3.20200123.0"
-    }
-}
